Fix login issues
