+++ authentication/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "passwordless-simple-mail",
+  "private": "false",
+  "version": "0.0.2",
+  "scripts": {
+    "start": "node app.js"
+  },
+  "dependencies": {
+    "body-parser": "~1.13.1",
+    "cookie-parser": "~1.3.5",
+    "ejs": "~2.3.1",
+    "emailjs": "^0.3.16",
+    "express": "~4.12.4",
+    "express-session": "^1.11.3",
+    "morgan": "~1.6.0",
+    "passwordless": "*",
+    "passwordless-mongostore": "*"
+  }
+}
