+++ web/public/pure-css-custom-form-elements/style.css
@@ -1,0 +1,319 @@
+/* These styles adapted from
+http://www.onextrapixel.com/2012/10/26/custom-style-all-your-form-elements-with-pure-css-and-no-javascript/
+*/
+
+
+/* ----------------------------------------------  Layout ----------------------------------------------  */
+
+body {
+  margin: 0;
+  padding: 0;
+}
+
+a {
+  color: #0181bc;
+}
+
+pre,
+code {
+  font-size: 0.75em;
+  color: black;
+  font-family: "Courier", "courier new", monospace;
+}
+
+code {
+  padding: 0.2em
+}
+
+pre {
+  padding: 1em;
+  white-space: pre-line;
+}
+
+
+/* ----------------------------------------------  Reset form elements ---------------------------------------------- */
+
+input,
+select,
+textarea {
+  margin: 0;
+  padding: 0;
+  font-size: 0.85em;
+  outline: none;
+  font-family: inherit;
+  -moz-box-sizing: border-box;
+  /* Firefox */
+  -webkit-box-sizing: border-box;
+  /* Safari */
+  box-sizing: border-box;
+}
+
+
+/* ----------------------------------------------  Inputs, textareas and selects  ---------------------------------------------- */
+
+input[type="text"],
+input[type="url"],
+input[type="password"],
+textarea,
+select,
+div.styled,
+input[type="file"] {
+  width: 24em;
+  border-radius: 2px;
+  border: solid 1px #ccc;
+  padding: 0.4em;
+}
+
+div.styled,
+select,
+input[type="submit"],
+input[type="button"],
+button,
+input[type="file"]:after {
+  background: white url(formelements-select.png) no-repeat center;
+  -webkit-box-shadow: 0 1px 3px rgba(0, 0, 0, 0.2);
+  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.2);
+}
+
+input[type="text"],
+input[type="url"],
+input[type="password"],
+textarea,
+input[type="file"] {
+  background-color: #f5f5f5;
+  -webkit-box-shadow: inset 0 2px 3px rgba(0, 0, 0, 0.2);
+  box-shadow: inset 0 2px 3px rgba(0, 0, 0, 0.2);
+}
+
+.ie9 input[type="text"],
+.ie9 input[type="url"],
+.ie9 input[type="password"]{
+  line-height: normal;
+}
+
+
+/* Get the stuff to line up right */
+
+textarea {
+  width: 100%;
+  height: 10em;
+}
+
+
+/* ----------------------------------------------  Select menu  ---------------------------------------------- */
+
+
+/* For IE and Firefox */
+
+div.styled {
+  overflow: hidden;
+  padding: 0;
+  margin: 0;
+}
+
+.ie7 div.styled {
+  border: none;
+}
+
+div.styled select {
+  width: 115%;
+  background-color: transparent;
+  background-image: none;
+  -webkit-appearance: none;
+  border: none;
+  box-shadow: none;
+}
+
+.ie7 div.styled select {
+  width: 100%;
+  background-color: #fff;
+  border: solid 1px #ccc;
+  padding: 0.3em 0.5em;
+}
+
+
+/* ----------------------------------------------  File field  ----------------------------------------------  */
+
+
+/* Webkit Only */
+
+input[type="file"] {
+  position: relative;
+  -webkit-appearance: none;
+  -webkit-box-sizing: border-box;
+  box-sizing: border-box;
+  padding: 0;
+}
+
+input[type=file]::-webkit-file-upload-button {
+  width: 0;
+  padding: 0;
+  margin: 0;
+  -webkit-appearance: none;
+  border: none;
+}
+
+input[type="file"]:after {
+  content: 'Browse...';
+  margin: 0 0 0 0.5em;
+  display: inline-block;
+  left: 100%;
+  position: relative;
+  background: white url(formelements-select.png) no-repeat center left;
+  padding: 0.55em 1em 0.5em 1em;
+  font-size: 0.8em;
+  border: solid 1px #ccc !important;
+  -webkit-box-shadow: 0 1px 3px rgba(0, 0, 0, 0.2);
+  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.2);
+  border-radius: 4px;
+}
+
+input[type="file"]:active:after {
+  box-shadow: none;
+}
+
+
+/* ----------------------------------------------  Checkboxes and Radio inputs  ----------------------------------------------  */
+
+input[type="radio"],
+input[type="checkbox"] {
+  position: relative;
+  left: -999em;
+  width: 25px;
+  height: 25px;
+}
+
+input[type="checkbox"]::before,
+input[type="radio"]::before {
+  display: inline-block;
+  position: relative;
+  top: -0.1em;
+  left: calc(999em + 2px);
+  content: '';
+  width: 25px;
+  height: 25px;
+  background-image: url(formelements.png);
+  cursor: pointer;
+}
+
+input[type="checkbox"]::before {
+  background-position: 0 -25px;
+}
+
+input[type="checkbox"]:checked::before {
+  background-position: 0 0;
+}
+
+input[type="checkbox"]:indeterminate::before {
+  background-position: 0 -50px;
+}
+
+input[type="radio"]::before {
+  background-position: -25px -25px;
+}
+
+input[type="radio"]:checked::before {
+  background-position: -25px 0;
+}
+
+input[type="checkbox"]:disabled::before,
+input[type="radio"]:disabled::before,
+label + input[type="checkbox"]:disabled,
+label + input[type="radio"]:disabled,
+input[type="checkbox"]:disabled + label,
+input[type="radio"]:disabled + label {
+  opacity: 0.5;
+  cursor: default;
+}
+
+
+/* Remove the custom styling for IE 7-8 */
+
+.ie8 label:before {
+  display: none;
+  content: none;
+}
+
+.ie8 input[type="checkbox"],
+.ie8 input[type="radio"],
+.ie7 input[type="checkbox"],
+.ie7 input[type="radio"] {
+  position: static;
+  left: 0;
+}
+
+.ie8 input[type="checkbox"],
+.ie8 input[type="radio"] {
+  position: relative;
+  top: 5px;
+  margin-right: 0.5em;
+}
+
+input[type="text"]:focus,
+textarea:focus {
+  border-color: #000;
+}
+
+
+/* ---------------------------------------------- Form Submit and Next buttons ---------------------------------------------- */
+
+input[type="submit"],
+input[type="button"],
+button {
+  padding: 0.5em 1em;
+  line-height: 1em;
+  cursor: pointer;
+  border-radius: 4px;
+  color: #000;
+  font-weight: bold;
+  font-size: inherit;
+  border: solid 1px #ccc;
+  box-shadow: 0 1px 5px rgba(0, 0, 0, 0.2);
+  background-position: center bottom;
+}
+
+
+/* Gradient Generator by : http://www.colorzilla.com/gradient-editor/ */
+
+input[type="submit"]:active,
+input[type="button"]:active,
+button:active {
+  -webkit-box-shadow: none;
+  box-shadow: none;
+}
+
+input[type="submit"]:disabled,
+input[type="button"]:disabled,
+button:disabled {
+  opacity: 0.5;
+  cursor: default;
+}
+
+
+/* IE7 needs you to change the background to transparent when using image background for submit buttons */
+
+
+/*----------------------------------------------  Media Queries ----------------------------------------------  */
+
+@media screen and (max-width: 600px) {
+  body {
+    width: 80%;
+    font-size: 15px;
+  }
+}
+
+
+/* end of query */
+
+@media screen and (max-width: 400px) {
+  input[type="text"],
+  input[type="password"],
+  textarea,
+  select,
+  div.styled {
+    width: 100%
+  }
+}
+
+
+/* end of query */
