+++ web/views/index.ejs
@@ -3,44 +3,46 @@
   <head>
     <link rel="stylesheet" type="text/css" href="style.css"/>
     <script type="application/javascript" src="https://d3js.org/d3.v4.min.js"></script>
+    <script type="application/javascript" src="https://code.jquery.com/jquery-3.1.1.min.js"></script>
+    <script type="application/javascript">
+      <% console.log(Object.keys(this)); %>
+      <% console.log(typeof serverState === 'undefined' ? 'undefined' : JSON.stringify(serverState, null, 4)); %>
+      <% if(typeof serverState === 'undefined') { var serverState = {}; } %>
+      <% if(serverState) { %>
+      	window.serverState = JSON.parse(jQuery('<textarea/>').html('<%= JSON.stringify(serverState) %>').text());
+      <% } else { %>
+        window.serverState = {};
+      <% } %>
+    </script>
     <script type="application/javascript" src="script.js"></script>
   </head>
   <body>
-    <div id="overview">
-      <div id="visualization">
-        TODO: cool visualization here
-      </div>
-      <div id="deviceStatusList">
-
-      </div>
+    <img src="logo.png" id="logo"/>
+    <div id="map">
+      <p class="tempMessage">TODO: put a map here?</p>
     </div>
-    <div id="userManagement">
-      <form id="logout">
-        <span>Logged in as <label id="currentUser"></label></span>
-        <button type="submit">Log out</button>
-      </form>
-      <form id="login">
-        <input type="text" placeholder="Username" name="uname" required/>
-        <input type="password" placeholder="Password" name="pwd" required/>
-        <button>Log in</button>
-      </form>
-      <form id="registerUser">
-        <input type="email" placeholder="Email Address" name="email" required/>
-        <input type="text" placeholder="Choose a username" name="uname" required/>
-        <input type="password" placeholder="Choose a password" name="pwd" required/>
-        <input type="password" placeholder="Re-enter your password" name="validate_pwd" required/>
-        <button>Register</button>
-      </form>
-      <form id="registerDevice">
-        <input type="text" name="mac_address" placeholder="Sensor MAC Address" required/>
-        <input type="number" name="latitude" placeholder="Latitude" required/>
-        <input type="number" name="longitude" placeholder="Longitude" required/>
-        <span>
-          <label>Location precision</label>
-          <input type="number" name="radius" required/>
-        </span>
-        <button>Register A Sensor</button>
-      </form>
+    <div id="sidebar">
+      <div id="hamburger">
+        <img src="hamburger.svg"/>
+      </div>
+      <div id="content">
+        <% if(serverState.user) { %>
+          <form id="logout" action="/logout">
+            <span>Logged in as <label id="currentUser"><%= serverState.user %></label></span>
+            <input type="submit" value="Log out"/>
+          </form>
+          <hr/>
+          <form id="registerDevice" action="/registerDevice" method="POST">
+            <input type="text" name="mac_address" placeholder="Sensor MAC Address" required/>
+            <input type="submit" value="Register Sensor"/>
+          </form>
+        <% } else { %>
+          <form id="login" action="/sendtoken" method="POST">
+            <input type="email" placeholder="Email Address" name="user" required/>
+            <input type="submit" value="Log in / Register"/>
+          </form>
+        <% } %>
+      </div>
     </div>
   </body>
 </html>
